add_air_yac_ep_variables {nflscrapR}	R Documentation

Calculate and add the air and yac expected points variables to include in a 'nflscrapR' play-by-play data frame

Description

Given a 'nflscrapR' play-by-play data frame, calculate the air and yac EPA for passing playsexpected points using the 'nflscrapR' expected points model and include append the column to the input data frame. See here for an explanation of the model methodology: https://arxiv.org/abs/1802.00998. Source code for fitting the model is located here https://github.com/ryurko/nflscrapR-models/blob/master/R/init_models/init_ep_fg_models.R.

Usage

add_air_yac_ep_variables(pbp_data)

Arguments

pbp_data
Data frame with all of the necessary columns used to estimate and include the air and yac epa

Value

The input data frame with additional columns included for the air EPA (air_epa), yac EPA (yac_epa), and cumulative totals for home and away teams (total_home_air_epa, total_home_yac_epa, etc.).

[Package nflscrapR version 1.8.0 Index]