agg_player_season {nflscrapR}	R Documentation

Detailed Player Aggregate Season Statistics

Description

This function outputs a dataframe with the season statistics for each player who recorded atleast one measured statistic in any game throughout the specified season. This function gives one line per player with the following statistics: Passing, Rushing, Receiving, Kick Return, Punt Return, Fumbles, and Defense

Usage

agg_player_season(Season, Weeks = 16)

Arguments

Season
(numeric) A 4-digit year associated with a given season

Details

This function calls season_player_game and then aggregates across an entire season to gather season totals and season max statistics

Value

Returns a dataframe with a single line for each player aggregating their total season statistics

Examples

# Returns the Season-Total Statistics for Each Player in the 2015 Season
agg_player_season(2015)

[Package nflscrapR version 1.8.0 Index]