buildNameAbbr {nflscrapR}	R Documentation

Build formatted player name from full player name

Description

This sub-function, called in the season_rosters function, takes the full name of each player and formats it into the first initial of their first name and last initial of their last name.

Usage

buildNameAbbr(value)

[Package nflscrapR version 1.8.0 Index]