create_game_html_url {nflscrapR}	R Documentation

Create the url with the location of raw NFL play-by-play HTML

Description

This function pastes creates the url for the provided game id. The returned url is used in the various play-by-play functions to access play-by-play data from the raw HTML page.

Usage

create_game_html_url(game_id)

Arguments

game_id
(character or numeric) A 10 digit game id associated with a given NFL game.

Value

String url where the raw HTML play-by-play data can be found (going back to 1998 for regular season, 2000 for pre season).

Examples

# Create the JSON url for the first game of the 1998 season:
create_game_html_url(2017090700) 

[Package nflscrapR version 1.8.0 Index]