nflteams {nflscrapR}	R Documentation

Dataset of NFL team names, abbreviations, and colors

Description

Dataset of NFL team names, abbreviations, and colors

Usage

nflteams

Format

A data frame with one row for each team/abbreviation combo that appears in the NFL play-by-play data with the following columns accessed from the 'teamcolors' package created by Ben Baumer and Gregory Matthews:

team

full name of the team

abbr

team abbreviation used in play-by-play data

primary

team's primary color

secondary

team's secondary color

tertiary

team's tertiary color

quaternary

team's quaternary color

division

team's division

Details

Team names, abbreviations, divisions, and team colors.

Examples

data(nflteams)

[Package nflscrapR version 1.8.0 Index]