scrape_game_ids {nflscrapR}	R Documentation

Scrape game ids for a given NFL season (either pre, regular, or post-season)

Description

This function returns all available game ids for a given season, with an option specifying which type of game ids to return (either the pre, regular, or post season) as well as the weeks considered.

Usage

scrape_game_ids(season, type = "reg", weeks = NULL, teams = NULL)

Arguments

season
Numeric 4-digit year associated with an NFL season

type
String indicating the type of game ids to find, must either be "pre", "reg", or "post" (default is "reg").

weeks
Numeric vector indicating which weeks of preseason (0 to 4 except for 2011 season without Hall of Fame game) or regular season games (1 to 17) to grab (default value of NULL corresponds to selecting all available weeks).

teams
String vector indicating which teams (based on the abbreviation) the function should grab game info for (default value of NULL corresponds to selecting all available teams).

Value

Data frame with columns containing the type, game_id, week number (if pre or regular season), and the year of the season. for each game id that is scraped.

Examples

# Scraping all game ids from 2017 regular season:
scrape_game_ids(2017) 

[Package nflscrapR version 1.8.0 Index]