season_games {nflscrapR}	R Documentation

Game Information for All Games in a Season

Description

This function intakes a year associated with a given season and outputs all the game matchups for all 17 weeks of the regular season

Usage

season_games(Season, sleep.seconds = 0)

Arguments

Season
(numeric): A 4-digit year associated with a given NFL season

sleep.seconds
(numeric): Allows the user to tell the function to sleep between calls to the API to avoid disrupting the connection. Note, this will make the function take much longer.

Week
(numeric): A number corresponding to the number of weeks of data you want to be scraped and included in the output.

Details

Reference the stored dataframe nflteams to match team abbreviations with the full team names

Value

A dataframe with the gameID, the game date, the home team abbreviation, and away team abbreviation

Examples

# All games in 2015 Season
season_games(2015) # Will output a dataframe

[Package nflscrapR version 1.8.0 Index]