season_rosters {nflscrapR}	R Documentation

Season Rosters for Teams

Description

This function intakes a year and a team abbreviation and outputs a dataframe with each player who has played for the specified team and recorded a measurable statistic

Usage

season_rosters(season, teams, positions = c("QUARTERBACK", "RUNNING_BACK",
  "WIDE_RECEIVER", "TIGHT_END"))

Arguments

season:
A 4-digit year associated with a given NFL season

teams:
A string vector containing the abbreviations for NFL Team(s)

Positions:
A string vector containing the abbreviations for NFL position(s). Can be of the following:

"QUARTERBACK" (in the default list)

"RUNNING_BACK" (in the default list)

"WIDE_RECEIVER" (in the default list)

"TIGHT_END" (in the default list)

"DEFENSIVE_LINEMAN"

"LINEBACKER"

"DEFENSIVE_BACK"

"KICKOFF_KICKER"

"KICK_RETURNER"

"PUNTER"

"PUNT_RETURNER"

"FIELD_GOAL_KICKER"

Details

To find team associated abbrevations use the nflteams dataframe stored in this package!

Value

A dataframe with columns associated with season/year, full player name, team initial, position, and formated player name. players who played and recorded some measurable statistic, and the last column specifyng the number of games they played in.

Examples

# Roster for Baltimore Ravens in 2013
season_rosters(2013, TeamInt = "BAL") 

[Package nflscrapR version 1.8.0 Index]