simple_boxscore {nflscrapR}	R Documentation

Simple Game Boxscore

Description

This function pulls data from an NFL url and contructs it into a formatted boxscore.

Usage

simple_boxscore(GameID, home = TRUE)

Arguments

GameID
(character or numeric) A 10 digit game ID associated with a given NFL game.

home
(boolean): home = TRUE will pull home stats, home = FALSE pulls away stats

Value

A list of player statistics including passing, rushing, receiving, defense, kicking, kick return, and punt return statistics for the specified game.

Examples

# Parsed drive summaries of final game in 2015 NFL season
nfl2015.finalregseasongame.gameID <- "2016010310"
simple_boxscore(nfl2015.finalregseasongame.gameID, home = TRUE) 

[Package nflscrapR version 1.8.0 Index]